# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Graph","description":"APIs for retrieving graph data"},{"name":"Positions","description":"APIs for managing item positions"},{"name":"Simulations","description":"APIs for creating and managing historical state simulations"}],"paths":{"/api/positions":{"put":{"tags":["Positions"],"summary":"Move an item to a new position","operationId":"moveConnection","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConnection"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["Positions"],"summary":"Create a new position connection","operationId":"createConnection","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConnection"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/locations":{"get":{"tags":["location-controller"],"operationId":"getAllLocations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Location"}}}}}}},"put":{"tags":["location-controller"],"operationId":"updateLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}},"post":{"tags":["location-controller"],"operationId":"createLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/items":{"get":{"tags":["item-controller"],"operationId":"getAllItems","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Item"}}}}}}},"put":{"tags":["item-controller"],"operationId":"updateItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}},"post":{"tags":["item-controller"],"operationId":"createItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/simulations":{"post":{"tags":["Simulations"],"summary":"Create a new simulation","description":"Initiates an asynchronous build of a historical graph state. Returns immediately with a simulation ID.","operationId":"createSimulation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSimulationRequest"}}},"required":true},"responses":{"202":{"description":"Simulation build has been accepted for processing.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimulationState"}}}}}}},"/api/simulations/{simulationId}/playback/start":{"post":{"tags":["Simulations"],"summary":"Start event playback","description":"Starts the time-synchronized event playback for a READY simulation.","operationId":"startPlayback","parameters":[{"name":"simulationId","in":"path","description":"The ID of the simulation","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaybackRequest"}}},"required":true},"responses":{"202":{"description":"Playback has been accepted and started."},"409":{"description":"Simulation is not in a READY state."}}}},"/api/simulations/{simulationId}/playback/cancel":{"post":{"tags":["Simulations"],"summary":"Cancel event playback","description":"Stops a currently running event playback.","operationId":"cancelPlayback","parameters":[{"name":"simulationId","in":"path","description":"The ID of the simulation","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Cancellation signal has been sent."}}}},"/api/simulations/{simulationId}/heartbeat":{"post":{"tags":["Simulations"],"summary":"Send a heartbeat","description":"A lightweight endpoint for the frontend to periodically call to keep a simulation alive.","operationId":"sendHeartbeat","parameters":[{"name":"simulationId","in":"path","description":"The ID of the active simulation","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Heartbeat received."}}}},"/api/connections":{"post":{"tags":["location-connection-controller"],"operationId":"createConnection_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectionInput"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["location-connection-controller"],"operationId":"deleteConnection","parameters":[{"name":"sourceId","in":"query","required":true,"schema":{"type":"string"}},{"name":"targetId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/simulations/{simulationId}":{"get":{"tags":["Simulations"],"summary":"Get simulation status","description":"Poll this endpoint to check the build status (e.g., QUEUED, BUILDING, READY).","operationId":"getSimulationStatus","parameters":[{"name":"simulationId","in":"path","description":"The unique ID of the simulation","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Simulation not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimulationState"}}}},"200":{"description":"Current state of the simulation.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimulationState"}}}}}},"delete":{"tags":["Simulations"],"summary":"Destroy a simulation","description":"Deletes an in-memory simulation and cancels any associated tasks.","operationId":"destroySimulation","parameters":[{"name":"simulationId","in":"path","description":"The unique ID of the simulation","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Simulation successfully destroyed."}}}},"/api/simulations/{simulationId}/graph":{"get":{"tags":["Simulations"],"summary":"Get graph data for a simulation","description":"Fetches the entire graph state for a specific simulation. Only works if the simulation status is READY or PLAYING.","operationId":"getSimulationGraphData","parameters":[{"name":"simulationId","in":"path","description":"The ID of the simulation","required":true,"schema":{"type":"string"}}],"responses":{"409":{"description":"Simulation is not in a READY state.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GraphData"}}}},"200":{"description":"The graph data for the simulation.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GraphData"}}}},"404":{"description":"Simulation not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GraphData"}}}}}}},"/api/locations/{id}":{"get":{"tags":["location-controller"],"operationId":"getLocationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Location"}}}}}},"delete":{"tags":["location-controller"],"operationId":"deleteLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/items/{id}":{"get":{"tags":["item-controller"],"operationId":"getItemById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Item"}}}}}},"delete":{"tags":["item-controller"],"operationId":"deleteItem","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/graph":{"get":{"tags":["Graph"],"summary":"Get the current state of the graph","operationId":"getGraphData","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GraphData"}}}}}}},"/api/positions/{itemId}":{"delete":{"tags":["Positions"],"summary":"Delete position connections for an item","operationId":"deleteConnections","parameters":[{"name":"itemId","in":"path","description":"ID of the item","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"CreateConnection":{"type":"object","properties":{"itemId":{"type":"string"},"locationId":{"type":"string"}}},"UpdateModel":{"type":"object","properties":{"id":{"type":"string"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"CreateSimulationRequest":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"}}},"SimulationState":{"type":"object","properties":{"id":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["QUEUED","BUILDING","READY","PLAYING","FAILED"]},"lastHeartbeatTimestamp":{"type":"string","format":"date-time"}}},"PlaybackRequest":{"type":"object","properties":{"speedFactor":{"type":"number","format":"double"}}},"LocationInput":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"length":{"type":"number","format":"double"},"speed":{"type":"number","format":"double"},"type":{"type":"string","enum":["CONVEYOR","JUNCTION","CHUTE","ACCUMULATION","ROAD","GENERIC"]},"capactiy":{"type":"integer","format":"int32"},"active":{"type":"boolean"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"ItemInput":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"speed":{"type":"number","format":"double"},"active":{"type":"boolean"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"ConnectionInput":{"type":"object","properties":{"location1Id":{"type":"string"},"location2Id":{"type":"string"}}},"ConnectionResponse":{"type":"object","properties":{"sourceId":{"type":"string"},"targetId":{"type":"string"},"direction":{"type":"string"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"GraphData":{"type":"object","properties":{"locations":{"type":"array","items":{"$ref":"#/components/schemas/LocationResponse"}},"connections":{"type":"array","items":{"$ref":"#/components/schemas/ConnectionResponse"}}}},"ItemResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"speed":{"type":"number","format":"double"},"active":{"type":"boolean"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"LocationResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"length":{"type":"number","format":"double"},"speed":{"type":"number","format":"double"},"type":{"type":"string","enum":["CONVEYOR","JUNCTION","CHUTE","ACCUMULATION","ROAD","GENERIC"]},"active":{"type":"boolean"},"capacity":{"type":"integer","format":"int32"},"properties":{"type":"object","additionalProperties":{"type":"object"}},"items":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"}},"connections":{"type":"array","items":{"$ref":"#/components/schemas/ConnectionResponse"}}}},"Location":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"type":{"type":"string","enum":["CONVEYOR","JUNCTION","CHUTE","ACCUMULATION","ROAD","GENERIC"]},"active":{"type":"boolean"},"properties":{"type":"object","additionalProperties":{"type":"object"}},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"length":{"type":"number","format":"double"},"speed":{"type":"number","format":"double"},"capacity":{"type":"integer","format":"int32"},"itemCount":{"type":"integer","format":"int32"},"outboundConnectionIds":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"full":{"type":"boolean"},"track":{"type":"boolean"},"transitTimeSeconds":{"type":"number","format":"double"}}},"Item":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"speed":{"type":"number","format":"double"},"active":{"type":"boolean"},"properties":{"type":"object","additionalProperties":{"type":"object"}},"location":{"$ref":"#/components/schemas/Location"},"progressInfo":{"$ref":"#/components/schemas/ProgressInfo"}}},"ProgressInfo":{"type":"object","properties":{"progress":{"type":"number","format":"double"},"datetime":{"type":"string","format":"date-time"}}}}}}